api/requirements.txt
bcrypt==4.2.1
Flask==3.0.3
mysql-connector-python==9.1.0
PyJWT==2.9.0